Move scribe into Note & Tasks; make dictation reusable on any text box
Address review feedback:
- Remove the separate "Scribe" encounter tab. The screening scribe now
  renders inside the Note & Tasks tab (ScribeTab -> ScribePanel).
- Add a reusable ScribeTextarea (Textarea + built-in dictation mic) so the
  voice-to-text scribe can live on any free text box. Applied to the chart
  note, the screening transcript, and the lang2fhir create page.
- useScribeTranscription is now callback-based and loads the Whisper model
  lazily on first mic click (not on mount), so rendering a scribe-enabled
  text box no longer downloads a model.
- Seed the chart note into local state once the ClinicalImpression loads
  (controlled ScribeTextarea no longer has defaultValue), guarded so it
  never clobbers in-progress edits.
- Tests updated: ScribePanel.test, EncounterChart.test (scoped chart-note
  textarea lookups; stub transformers).
