+++ src/components/encounter/EncounterHeader.tsx
@@ -200,7 +200,6 @@ export const EncounterHeader = (props: EncounterHeaderProps): JSX.Element => {
             onChange={handleTabChange}
             data={[
               { label: 'Note & Tasks', value: 'notes' },
-              { label: 'Scribe', value: 'scribe' },
               { label: 'Details & Billing', value: 'details' },
             ]}
             fullWidth
