+++ src/pages/resource/ResourceLang2FHIRCreatePage.tsx
@@ -1,6 +1,6 @@
 // SPDX-FileCopyrightText: Copyright Orangebot, Inc. and Medplum contributors
 // SPDX-License-Identifier: Apache-2.0
-import { Stack, Text, Textarea, Button, Box, Space, ActionIcon, Group } from '@mantine/core';
+import { Stack, Text, Button, Box, Space } from '@mantine/core';
 import { showNotification } from '@mantine/notifications';
 import { normalizeErrorString, normalizeOperationOutcome } from '@medplum/core';
 import type { OperationOutcome, Resource, ResourceType } from '@medplum/fhirtypes';
@@ -9,10 +9,10 @@ import type { JSX } from 'react';
 import { useState, useEffect } from 'react';
 import { useNavigate, useParams } from 'react-router';
 import { usePatient } from '../../hooks/usePatient';
-import { useScribeTranscription } from '../../hooks/useScribeTranscription';
+import { ScribeTextarea } from '../../components/ScribeTextarea';
 import { prependPatientPath } from '../patient/PatientPage.utils';
 import { PhenoMLBranding } from '../../components/PhenoMLBranding';
-import { IconSparkles, IconMicrophone, IconMicrophoneOff } from '@tabler/icons-react';
+import { IconSparkles } from '@tabler/icons-react';
 
 
 // Define which resource types don't require a patient
@@ -22,15 +22,7 @@ type PatientIndependentResource = typeof PATIENT_INDEPENDENT_RESOURCES[number];
 export function ResourceLang2FHIRCreatePage(): JSX.Element {
   const medplum = useMedplum();
   const [outcome, setOutcome] = useState<OperationOutcome | undefined>();
-  const {
-    transcript: inputText,
-    setTranscript: setInputText,
-    isRecording,
-    isModelLoading,
-    isProcessing,
-    startRecording,
-    stopRecording,
-  } = useScribeTranscription();
+  const [inputText, setInputText] = useState('');
   const navigate = useNavigate();
   const { patientId, resourceType } = useParams() as { patientId: string | undefined; resourceType: ResourceType };
   const [loading, setLoading] = useState(false);
@@ -55,11 +47,7 @@ export function ResourceLang2FHIRCreatePage(): JSX.Element {
       if (outcome) {
         setOutcome(undefined);
       }
-      
-      if (isRecording) {
-        stopRecording();
-      }
-      
+
       setLoading(true);
       const lang2fhirCreateBot = await medplum.searchOne('Bot', { name: 'lang2fhir-create' });
       if (!lang2fhirCreateBot?.id) {
@@ -111,38 +99,18 @@ export function ResourceLang2FHIRCreatePage(): JSX.Element {
       <Stack>
         <Text fw={500}>Create a new {resourceType} using Natural Language</Text>
         
-        <Group align="flex-start">
-          <Textarea
-            style={{ flex: 1 }}
-            label="Enter your description"
-            placeholder={`Describe the ${resourceType.toLowerCase()} in natural language...`}
-            minRows={4}
-            value={inputText}
-            onChange={(e) => setInputText(e.currentTarget.value)}
-          />
-          <ActionIcon 
-            size="lg"
-            variant="light"
-            color={isRecording ? "red" : "blue"}
-            onClick={isRecording ? stopRecording : startRecording}
-            mt={25}
-            disabled={isModelLoading || isProcessing}
-            title={isRecording ? "Stop recording" : "Start recording"}
-          >
-            {isRecording ? <IconMicrophoneOff size={20} /> : <IconMicrophone size={20} />}
-          </ActionIcon>
-        </Group>
+        <ScribeTextarea
+          label="Enter your description"
+          placeholder={`Describe the ${resourceType.toLowerCase()} in natural language...`}
+          minRows={4}
+          value={inputText}
+          onChange={setInputText}
+        />
 
-        {(isModelLoading || isProcessing) && (
-          <Text size="sm" c="dimmed">
-            {isModelLoading ? "Loading speech recognition model..." : "Processing audio..."}
-          </Text>
-        )}
-
-        <Button 
-          onClick={handleSubmit} 
+        <Button
+          onClick={handleSubmit}
           loading={loading}
-          disabled={!inputText.trim() || isRecording || isProcessing}
+          disabled={!inputText.trim()}
         >
           <IconSparkles size={14} style={{ marginRight: 8 }} />
           Create {resourceType}
